use innerHTML instead ngModel
We tried to use your ngModel (which is not documented) but it doesn't work we got some errors on, so we changed it to innerHTML input property which works well. Also we add a delay property to setFontSize which we set to 0 if setFontSize is called on angular's live cycles. Only on window:resize we keep the default or passed delay time.